% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfilogisreg.R
\name{gfilogisreg}
\alias{gfilogisreg}
\title{Generalized fiducial inference for logistic regression}
\usage{
gfilogisreg(formula, data = NULL, N, thresh = N/2)
}
\arguments{
\item{formula}{formula describing the model}

\item{data}{dataframe containing the variables in the model}

\item{N}{number of fiducial simulations}

\item{thresh}{threshold criterion for the alteration; expert usage only}
}
\value{
A list with two fields: \code{Beta}, the fiducial simulations of
  the parameters, and \code{Weights}, their weight.
}
\description{
Simulates the fiducial distribution of a logistic regression
  model.
}
\examples{
y <- c(
  0, 0, 0, 1,
  0, 0, 1, 1,
  0, 1, 1, 1
)
group <- gl(3, 4)
gf <- gfilogisreg(y ~ 0 + group, N = 1000)
}
