% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{gfiSummary}
\alias{gfiSummary}
\title{Summary of fiducial samples}
\usage{
gfiSummary(fidsamples, conf = 0.95)
}
\arguments{
\item{fidsamples}{fiducial samples, the output of \code{\link{gfilogisreg}}}

\item{conf}{confidence level}
}
\value{
A matrix with summary statistics: means, medians, and confidence
  intervals.
}
\description{
Summary of the fiducial samples.
}
\examples{
y <- c(
  0, 0, 0, 1,
  0, 0, 1, 1,
  0, 1, 1, 1
)
group <- gl(3, 4)
fidsamples <- gfilogisreg(y ~ 0 + group, N = 1000)
gfiSummary(fidsamples)
}
